[release-branch.go1.2] blog: Fix atom feed's `updated' time when there's only one article.
««« CL 18420044 / e145aeb0df07
blog: Fix atom feed's `updated' time when there's only one article.

When computing the time for the "updated" tag of the atom feed, the
current code checks if there is more than one article and if that is
not true, it sets the time to the zero time.Time. This means that
the feed also gets the zero time in this tag when there is exactly one
article.

This trivial patch fixes this so that when there is exactly one
article, the time is set to that article's time.

R=golang-dev, adg
CC=golang-dev
https://golang.org/cl/18420044

»»»

R=golang-dev, dsymonds
CC=golang-dev
https://golang.org/cl/23390043