cmd/present: update the instructions for App Engine
The command's documentation mistakingly said that the instructions for
testing and deployment to App Engine were contained in the top-level
README for the tools repository.  As best as I can tell, the top-level
README never contained this information.

This commit adds a configuration fragment and provides sample
instructions for the App Engine workflow.

Change-Id: If89c099e254ef91ca224742aebd81f7d1206d9b8
Reviewed-on: https://go-review.googlesource.com/15685
Reviewed-by: Andrew Gerrand <adg@golang.org>