cmd/cover: show uncovered count in HTML dropdown
The HTML reports as implemented currently show a percentage coverage
for each file in the file selector. This is not particularly useful
when you have a large number of files and are hunting for which ones
are lowering your overall coverage. A file with 66% coverage but only a
few lines is not nearly as suspect as a file with the same percentage
but thousands of lines.

This modifies the HTML report to also show a count of uncovered lines
next to the percentage, so that it's easier to spot files that might
need more testing attention in a large project.