present: add footer template This commit add the possibily to define a footer template using two lines in the header section.
The footer line is a pipe (|) list of three segments that can be used to add a footer
for each slide. The justifications are left, center, and right.
Use [#] to denote the slide number and .image to add an image.

The footer omit line is a comma-separated list of slide numbers that can be used
to turn off the footer for a specific slide.