present: remove mention of non-existing emphasis toggle
To find out whether the documentation was wrong, or if the
implementation had a bug, I did some code archeology. The earliest
commit I found where the 'h' key in browser was mentioned was the
very first commit where the present format is added. It was in talks
subrepo before being moved to tools subrepo. It was CL 6497063.

In that commit, I see no mention of 'h' key anywhere else except
that one line of documentation.

Three years later, the 'h' key got mapped to hiding the help dialog
in CL 4910.

My best guess is this original feature was documented but
never implemented. So removing it from documentation is
the most appropriate fix.

Fixes golang/go#17375.

Change-Id: Ibe0b39d73a7a0652acd6a04beddfcff22b0e3c4a
Reviewed-on: https://go-review.googlesource.com/c/143557
Reviewed-by: Andrew Gerrand <adg@golang.org>